LayoutTests/imported/w3c/web-platform-tests/webcodecs/video-encoder-rescaling.https.any.worker_h264_avc-expected.txt


FAIL Scaling Image in Encoding from 64 x 64 to 128 x 128 Format: RGBX assert_approx_equals: Scaled Image differs too much! Scaling from 64 x 64 to 128 x 128 Format:RGBX expected 0 +/- 0.05 but got 1.287769778050109
FAIL Scaling Image in Encoding from 64 x 64 to 128 x 128 Format: I420 assert_approx_equals: Scaled Image differs too much! Scaling from 64 x 64 to 128 x 128 Format:I420 expected 0 +/- 0.05 but got 1.2668002791394335
FAIL Scaling Image in Encoding from 128 x 128 to 128 x 128 Format: RGBX assert_approx_equals: Scaled Image differs too much! Scaling from 128 x 128 to 128 x 128 Format:RGBX expected 0 +/- 0.05 but got 1.287840413943355
FAIL Scaling Image in Encoding from 128 x 128 to 128 x 128 Format: I420 assert_approx_equals: Scaled Image differs too much! Scaling from 128 x 128 to 128 x 128 Format:I420 expected 0 +/- 0.05 but got 1.2669798474945535
FAIL Scaling Image in Encoding from 128 x 128 to 64 x 64 Format: RGBX assert_approx_equals: Scaled Image differs too much! Scaling from 128 x 128 to 64 x 64 Format:RGBX expected 0 +/- 0.05 but got 1.28804891748366
FAIL Scaling Image in Encoding from 128 x 128 to 64 x 64 Format: I420 assert_approx_equals: Scaled Image differs too much! Scaling from 128 x 128 to 64 x 64 Format:I420 expected 0 +/- 0.05 but got 1.267391748366013
FAIL Scaling Image in Encoding from 32 x 32 to 96 x 96 Format: RGBX assert_approx_equals: Scaled Image differs too much! Scaling from 32 x 32 to 96 x 96 Format:RGBX expected 0 +/- 0.05 but got 1.287140674170903
FAIL Scaling Image in Encoding from 32 x 32 to 96 x 96 Format: I420 assert_approx_equals: Scaled Image differs too much! Scaling from 32 x 32 to 96 x 96 Format:I420 expected 0 +/- 0.05 but got 1.266417029774873
FAIL Scaling Image in Encoding from 192 x 192 to 64 x 64 Format: RGBX assert_approx_equals: Scaled Image differs too much! Scaling from 192 x 192 to 64 x 64 Format:RGBX expected 0 +/- 0.05 but got 1.2880718954248365
FAIL Scaling Image in Encoding from 192 x 192 to 64 x 64 Format: I420 assert_approx_equals: Scaled Image differs too much! Scaling from 192 x 192 to 64 x 64 Format:I420 expected 0 +/- 0.05 but got 1.2675245098039216
FAIL Scaling Image in Encoding from 64 x 32 to 128 x 64 Format: RGBX assert_approx_equals: Scaled Image differs too much! Scaling from 64 x 32 to 128 x 64 Format:RGBX expected 0 +/- 0.05 but got 1.2883952886710242
FAIL Scaling Image in Encoding from 64 x 32 to 128 x 64 Format: I420 assert_approx_equals: Scaled Image differs too much! Scaling from 64 x 32 to 128 x 64 Format:I420 expected 0 +/- 0.05 but got 1.2674521718409586
FAIL Scaling Image in Encoding from 128 x 256 to 64 x 128 Format: RGBX assert_approx_equals: Scaled Image differs too much! Scaling from 128 x 256 to 64 x 128 Format:RGBX expected 0 +/- 0.05 but got 1.2873234102668845
FAIL Scaling Image in Encoding from 128 x 256 to 64 x 128 Format: I420 assert_approx_equals: Scaled Image differs too much! Scaling from 128 x 256 to 64 x 128 Format:I420 expected 0 +/- 0.05 but got 1.266689644607843
FAIL Scaling Image in Encoding from 64 x 64 to 128 x 192 Format: RGBX assert_approx_equals: Scaled Image differs too much! Scaling from 64 x 64 to 128 x 192 Format:RGBX expected 0 +/- 0.05 but got 1.2875873729121279
FAIL Scaling Image in Encoding from 64 x 64 to 128 x 192 Format: I420 assert_approx_equals: Scaled Image differs too much! Scaling from 64 x 64 to 128 x 192 Format:I420 expected 0 +/- 0.05 but got 1.2666051084785765
FAIL Scaling Image in Encoding from 128 x 192 to 64 x 64 Format: RGBX assert_approx_equals: Scaled Image differs too much! Scaling from 128 x 192 to 64 x 64 Format:RGBX expected 0 +/- 0.05 but got 1.2880804057734205
FAIL Scaling Image in Encoding from 128 x 192 to 64 x 64 Format: I420 assert_approx_equals: Scaled Image differs too much! Scaling from 128 x 192 to 64 x 64 Format:I420 expected 0 +/- 0.05 but got 1.267841094771242
